Move code to "wrappedTransaction.go".
